Move (almost) all tailwind styles to css file
To simplify the code examples in the post, move all of the tailwind
classes into globals.css. This will allow the post to be focused on
functionality, leaving all the styling to be baked in to the project
repo, only discoverable if the user is interested.

There are a couple places where styling is toggled via some condition
and needs to stay in the component.